requirements.txt
setproctitle
attrdict
librosa
pydub
pyloudnorm
x-transformers
torchdiffeq
openai-whisper>=20240930
transformers==4.48.3